requirements.txt
langchain>=0.1.0
langgraph>=0.0.15
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.5.0
typing-extensions>=4.8.0
python-twitter-v2>=0.8.1
google-api-python-client>=2.108.0
streamlit>=1.31.0
langchain_openai>=0.0.5